requirements.txt
gammapy==1.0.1
ipython==8.7.0
cython
jupyter
jupyterlab==3.5.3
pyyaml==6.0
click==8.1.3
numpy==1.24.2
scipy==1.10.1
pandas==1.5.2
matplotlib==3.7.1
astropy==5.1
regions==0.7
iminuit==2.20
sherpa==4.15.0
healpy==1.16.1
naima==0.10.0
pydantic==1.9.0
emcee==3.1.3
corner==2.2.1
requests
tqdm
coverage~=7.2.5
pytest~=7.3.1
pytest-cov~=4.0.0
unittest-xml-reporting~=3.2.0